Revert "feat: share glass capsule container helpers"
